Crate servo_rand

Structs

OS_RNG

Seed

ServoReseeder

ServoRng

ServoThreadRng

Constants

RESEED_THRESHOLD

SERVO_THREAD_RNG

Traits

Rng

An automatically-implemented extension trait on RngCore providing high-level generic methods for sampling values and other convenience methods.

RngCore

The core of a random number generator.

SeedableRng

A random number generator that can be explicitly seeded.

SliceRandom

Extension trait on slices, providing random mutation and sampling methods.

Functions

random

random_uuid

thread_rng