Crate servo_rand

Structs

ServoRng

ServoThreadRng

Traits

Rand
A type that can be randomly generated using an Rng.

Rng
A random number generator.

SeedableRng
A random number generator that can be explicitly seeded to produce the same stream of randomness multiple times.

Functions

random

thread_rng